refactor(web): migrate ProductRegistrationForm to new form conventions
- Move to product/registration-form/ subdirectory
- Rename ProductRegistrationForm.tsx to Form.tsx
- Create fields.ts with types, defaults, and validation
- Rename registrationFormOptions to defaultOptions
- Update import in ProductRegistrationPage
